ai_models_server/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
pillow
torch
transformers==4.51.3
accelerate
qwen-vl-utils
